Split to new file
